fix(mcp): get_skill installable:false for quarantined skills + run the SSRF e2e suite in CI (SMI-5360 Wave 5 PR1)
Part A (MCP twin, app code): executeGetSkill recomputed installable from
repo_url alone on both the API path (get-skill.ts) and the local-DB fallback,
so a quarantined skill was reported installable:true next to its security
warning — a self-contradictory response (install_skill already refuses
quarantined skills via validate.ts:149). Now:
  - API path: installable = Boolean(repo_url) && quarantined !== true.
  - Local path: consult QuarantineRepository.isQuarantined(id) (local quarantine
    lives in a separate table, not on the skill row) like the install path does.
  - formatSkillDetails distinguishes a quarantine block (has repo, not
    installable -> "blocked") from a genuine discovery-only entry (no repo).
Extracted formatSkillDetails + its trust-tier/score-bar helpers into
get-skill.format.ts (re-exported) to keep get-skill.ts under the 500-line limit.

Part B (CI lane, infra): packages/core/tests/e2e/security/security.e2e.test.ts
(SSRF / DNS-rebinding, offline) ran nowhere — excluded from the main run and
absent from both e2e configs. Added a test:e2e:security npm script
(vitest --config vitest-e2e.config.ts, scoped by positional path), added the
core e2e glob to that shared config's include (the CLI/MCP runners pass their
own dir positional, so their runs are unaffected), and a dedicated
security-e2e-tests job in e2e-tests.yml with an explicit numTotalTests>0
assertion so an inert lane cannot read as a pass.

Tests: API-path quarantined/clean/discovery-only installable branches; local-DB
quarantined seeded skill -> installable:false + "blocked"; formatter branches.
ADR-109: plan-reviewed (docs/internal/implementation/smi-5360-wave5-...).
Edge skills-get half ships separately (PR2, staging-gated).
