+++ vitest-e2e.config.ts
@@ -21,6 +21,11 @@ export default defineConfig({
       // MCP server E2E tests
       'packages/mcp-server/tests/e2e/**/*.test.ts',
       'packages/mcp-server/tests/e2e/**/*.e2e.test.ts',
+      // SMI-5360: core E2E security suite (SSRF / DNS-rebinding). Run only by the
+      // `test:e2e:security` script, which scopes via a positional path. The CLI/
+      // MCP runners pass their own dir positional (CLI_TESTS_DIR / MCP_TESTS_DIR),
+      // so this entry never widens their runs.
+      'packages/core/tests/e2e/**/*.e2e.test.ts',
     ],
     exclude: ['**/node_modules/**', '**/dist/**'],
     testTimeout: 60000, // 60s for E2E (overrides preset 15s)
